feat: add user profile management and update user information endpoints

+++ src/utils/fileremove.js
@@ -1,4 +1,6 @@
 import fs from "fs";
 export function fileRemove(localfilePath) {
+
+  if(fs.existsSync(localfilePath))
   fs.unlinkSync(localfilePath);
 }
